+++ src/world/roomJsonReaderWriter.py
@@ -6,6 +6,7 @@ import jsonschema
 from config.config import Config
 from entity.apple import Apple
 from entity.banana import Banana
+from entity.bed import Bed
 from entity.coalOre import CoalOre
 from entity.food import Food
 from entity.grass import Grass
@@ -17,6 +18,7 @@ from entity.living.chicken import Chicken
 from entity.living.livingEntity import LivingEntity
 from entity.oakWood import OakWood
 from entity.stone import Stone
+from entity.woodFloor import WoodFloor
 from lib.graphik.src.graphik import Graphik
 from lib.pyenvlib.grid import Grid
 from lib.pyenvlib.location import Location
@@ -208,6 +210,12 @@ class RoomJsonReaderWriter:
             entity.setID(UUID(entityJson["id"]))
         elif entityClass == "Banana":
             entity = Banana()
+            entity.setID(UUID(entityJson["id"]))
+        elif entityClass == "WoodFloor":
+            entity = WoodFloor()
+            entity.setID(UUID(entityJson["id"]))
+        elif entityClass == "Bed":
+            entity = Bed()
             entity.setID(UUID(entityJson["id"]))
         elif entityClass == "Player":
             return None